% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{fishdata_connect}
\alias{fishdata_connect}
\title{Connect to the Fishdb database

Uses connection parameters from \code{fishdb_options()}}
\usage{
fishdata_connect()
}
\value{
a connection object
}
\description{
Connect to the Fishdb database

Uses connection parameters from \code{fishdb_options()}
}
\examples{
\dontrun{
con <- fishdata_connect()
}

}
