% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{fishdata_disconnect}
\alias{fishdata_disconnect}
\title{Disconnect from FishDB}
\usage{
fishdata_disconnect(con)
}
\arguments{
\item{con}{}
}
\value{
TRUE invisibly.
}
\description{
Disconnect from FishDB
}
\examples{
fishdata_disconnect(con)

}
