% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{fishdb_disconnect}
\alias{fishdb_disconnect}
\title{Disconnect from FishDB}
\usage{
fishdb_disconnect(con)
}
\arguments{
\item{con}{}
}
\value{
fishdata_disconnect() retuns TRUE invisibly.
}
\description{
Disconnect from FishDB
}
\examples{
fishdb_disconnect(con)

}
