% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPTIONS.R
\name{fishdb_options}
\alias{fishdb_options}
\title{Set or get options for reading APEX Spöreg data}
\usage{
fishdb_options(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\description{
Set or get options for reading APEX Spöreg data
}
\section{Supported options}{

The following options are supported
\itemize{
\item{\code{dbname} Name of database (Default: fishdb)}
\item{\code{host} Name or ip-number (Default: 193.10.96.167)}
\item{\code{port} Port-number (Default: 5432)}
\item{\code{user} Username in database (Default: readonly)}
\item{\code{password} Password for user (Default: readonly)}
\item{\code{use_latin1} Switch client to latin1 (Default: FALSE)}
}
}

